GridManager is now more versatile and will instantiate any piece, piecemenu is as versatile now aswell.

+++ GMD200-Group-GridGame-ChessFight/Assets/Scripts/Pawn.cs
@@ -8,7 +8,7 @@ public class Pawn : MonoBehaviour
     [SerializeField] private GridManager gridManager;
     [SerializeField] private GridTile gridTile;
     //these will be set when the tower is instantiated so it is dynamic
-    public int xPos, yPos;
+    public float xPos, yPos;
 
     // Update is called once per frame
     void Update()
